+++ ember-file-upload/src/index.ts
@@ -1,11 +1,8 @@
-import UploadFile from './upload-file';
 import Queue from './queue';
-import HTTPRequest from './system/http-request';
-import DataTransferWrapper from './system/data-transfer-wrapper';
-import UploadFileReader from './system/upload-file-reader';
+import UploadFile from './upload-file';
+import uploadHandler from './mirage/upload-handler';
 import { FileSource, FileState } from './interfaces';
 import { DEFAULT_QUEUE } from './services/file-queue';
-import uploadHandler from './mirage/upload-handler';
 
 export {
   // Public API classes
@@ -13,10 +10,6 @@ export {
   UploadFile,
   // Mirage handler
   uploadHandler,
-  // Undocumented classes
-  HTTPRequest,
-  DataTransferWrapper,
-  UploadFileReader,
   // Enums
   FileSource,
   FileState,
