+++ ember-file-upload/src/internal.ts
@@ -1,0 +1,10 @@
+import DataTransferWrapper from './system/data-transfer-wrapper';
+import HTTPRequest from './system/http-request';
+import UploadFileReader from './system/upload-file-reader';
+
+export {
+  // Non-public classes imported by the test app
+  DataTransferWrapper,
+  HTTPRequest,
+  UploadFileReader,
+};
