+++ test-app/tests/unit/system/upload-file-reader-test.js
@@ -1,4 +1,4 @@
-import { UploadFileReader } from 'ember-file-upload';
+import { UploadFileReader } from 'ember-file-upload/internal';
 import { module, test } from 'qunit';
 
 const FakeFileReader = {
